requirements.txt
# requirements
pandas
numpy

# local package
-e .

# external requirements
coverage
python-dotenv
black